fix: handle test with new layout
