+++ cypress/support/commands.ts
@@ -1,5 +1,6 @@
 /// <reference types="cypress" />
 
+// eslint-disable-next-line @typescript-eslint/no-namespace
 declare namespace Cypress {
   interface Chainable {
     /**
