+++ cypress/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "compilerOptions": {
-    "target": "ES5",
-    "lib": ["ES6", "DOM"],
-    "types": ["cypress"]
+    "target": "es5",
+    "lib": ["es5", "dom"],
+    "types": ["cypress", "node"]
   },
   "include": ["**/*.ts"]
 }
